# internal-api/docs/swagger.yaml
basePath: /api/v1
definitions:
  api.CreateFactResult:
    properties:
      id:
        type: string
    type: object
  api.CreateUpdateFact:
    properties:
      fact:
        type: string
      source:
        type: string
    type: object
  api.ErrorResult:
    properties:
      error:
        type: string
    type: object
  repository.Fact:
    properties:
      approved:
        type: boolean
      createdAt:
        type: string
      createdBy:
        type: string
      fact:
        type: string
      id:
        type: string
      source:
        type: string
      updatedAt:
        type: string
      updatedBy:
        type: string
    type: object
externalDocs:
  description: OpenAPI
  url: https://swagger.io/resources/open-api/
host: https://animal-facts-internal.cafo.dev
info:
  contact: {}
  description: This API provides facts about animals.
  license:
    name: MIT
    url: https://github.com/cafo13/animal-facts/blob/main/LICENSE
  title: Animal Facts Internal API
  version: 0.0.4
paths:
  /facts:
    post:
      description: create a new fact
      parameters:
      - description: fact
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.CreateUpdateFact'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/api.CreateFactResult'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResult'
      summary: create fact
  /facts/:id:
    delete:
      description: delete an existing fact
      produces:
      - application/json
      responses:
        "200":
          description: fact deleted
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResult'
      summary: delete fact
    put:
      description: update an existing fact
      parameters:
      - description: fact
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.CreateUpdateFact'
      produces:
      - application/json
      responses:
        "200":
          description: fact updated
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResult'
      summary: update fact
  /facts/:id/approve:
    post:
      description: approve an existing fact, so that it gets available in the public
        API
      produces:
      - application/json
      responses:
        "200":
          description: fact approved
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResult'
      summary: approve fact
  /facts/:id/unapprove:
    post:
      description: unapprove an existing fact, so that it is no longer available in
        the public API
      produces:
      - application/json
      responses:
        "200":
          description: fact unapproved
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResult'
      summary: unapprove fact
  /facts/all:
    get:
      description: gets all facts (approved and unapproved) from the database
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              items:
                $ref: '#/definitions/repository.Fact'
              type: array
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResult'
      summary: gets all facts
swagger: "2.0"